move publicapi service from backend directory
