+++ manifests/bucketeer/charts/backend/templates/deployment.yaml
@@ -114,6 +114,8 @@ spec:
               value: "{{ .Values.env.auditLogServicePort }}"
             - name: BUCKETEER_BACKEND_AUTO_OPS_SERVICE_PORT
               value: "{{ .Values.env.autoOpsServicePort }}"
+            - name: BUCKETEER_BACKEND_PUBLIC_API_SERVICE_PORT
+              value: "{{ .Values.env.publicApiServicePort }}"
             - name: BUCKETEER_BACKEND_ENVIRONMENT_SERVICE_PORT
               value: "{{ .Values.env.environmentServicePort }}"
             - name: BUCKETEER_BACKEND_EVENT_COUNTER_SERVICE_PORT
