+++ pkg/backend/cmd/server/server.go
@@ -31,7 +31,6 @@ import (
 	"github.com/bucketeer-io/bucketeer/pkg/auth/oidc"
 	autoopsapi "github.com/bucketeer-io/bucketeer/pkg/autoops/api"
 	autoopsclient "github.com/bucketeer-io/bucketeer/pkg/autoops/client"
-	publicapi "github.com/bucketeer-io/bucketeer/pkg/backend/api"
 	btclient "github.com/bucketeer-io/bucketeer/pkg/batch/client"
 	"github.com/bucketeer-io/bucketeer/pkg/cache"
 	cachev3 "github.com/bucketeer-io/bucketeer/pkg/cache/v3"
@@ -49,6 +48,7 @@ import (
 	migratemysqlapi "github.com/bucketeer-io/bucketeer/pkg/migration/mysql/api"
 	"github.com/bucketeer-io/bucketeer/pkg/migration/mysql/migrate"
 	notificationapi "github.com/bucketeer-io/bucketeer/pkg/notification/api"
+	publicapi "github.com/bucketeer-io/bucketeer/pkg/publicapi/api"
 	"github.com/bucketeer-io/bucketeer/pkg/pubsub"
 	"github.com/bucketeer-io/bucketeer/pkg/pubsub/publisher"
 	pushapi "github.com/bucketeer-io/bucketeer/pkg/push/api"
