+++ proto/proto.lock
@@ -3127,117 +3127,6 @@
       }
     },
     {
-      "protopath": "backend:/:service.proto",
-      "def": {
-        "messages": [
-          {
-            "name": "GetFeatureRequest",
-            "fields": [
-              {
-                "id": 1,
-                "name": "environment_id",
-                "type": "string"
-              },
-              {
-                "id": 2,
-                "name": "id",
-                "type": "string"
-              }
-            ]
-          },
-          {
-            "name": "GetFeatureResponse",
-            "fields": [
-              {
-                "id": 1,
-                "name": "feature",
-                "type": "feature.Feature"
-              }
-            ]
-          },
-          {
-            "name": "UpdateFeatureRequest",
-            "fields": [
-              {
-                "id": 1,
-                "name": "comment",
-                "type": "string"
-              },
-              {
-                "id": 2,
-                "name": "environment_id",
-                "type": "string"
-              },
-              {
-                "id": 3,
-                "name": "feature",
-                "type": "feature.Feature"
-              }
-            ]
-          },
-          {
-            "name": "UpdateFeatureResponse"
-          }
-        ],
-        "services": [
-          {
-            "name": "PublicAPIService",
-            "rpcs": [
-              {
-                "name": "GetFeature",
-                "in_type": "GetFeatureRequest",
-                "out_type": "GetFeatureResponse",
-                "options": [
-                  {
-                    "name": "(google.api.http)",
-                    "aggregated": [
-                      {
-                        "name": "get",
-                        "value": "environment/{environment-id}/features/{feature-id}"
-                      }
-                    ]
-                  }
-                ]
-              },
-              {
-                "name": "UpdateFeature",
-                "in_type": "UpdateFeatureRequest",
-                "out_type": "UpdateFeatureResponse",
-                "options": [
-                  {
-                    "name": "(google.api.http)",
-                    "aggregated": [
-                      {
-                        "name": "patch",
-                        "value": "environment/{environment-id}/features/{feature-id}"
-                      }
-                    ]
-                  }
-                ]
-              }
-            ]
-          }
-        ],
-        "imports": [
-          {
-            "path": "google/api/annotations.proto"
-          },
-          {
-            "path": "proto/feature/feature.proto"
-          }
-        ],
-        "package": {
-          "name": "bucketeer.backend"
-        },
-        "options": [
-          {
-            "name": "go_package",
-            "value": "github.com/bucketeer-io/bucketeer/proto/backend"
-          }
-        ]
-      }
-    },
-    {
       "protopath": "batch:/:service.proto",
       "def": {
         "enums": [
@@ -16969,6 +16858,117 @@
           {
             "name": "go_package",
             "value": "github.com/bucketeer-io/bucketeer/proto/notification"
+          }
+        ]
+      }
+    },
+    {
+      "protopath": "publicapi:/:service.proto",
+      "def": {
+        "messages": [
+          {
+            "name": "GetFeatureRequest",
+            "fields": [
+              {
+                "id": 1,
+                "name": "environment_id",
+                "type": "string"
+              },
+              {
+                "id": 2,
+                "name": "id",
+                "type": "string"
+              }
+            ]
+          },
+          {
+            "name": "GetFeatureResponse",
+            "fields": [
+              {
+                "id": 1,
+                "name": "feature",
+                "type": "feature.Feature"
+              }
+            ]
+          },
+          {
+            "name": "UpdateFeatureRequest",
+            "fields": [
+              {
+                "id": 1,
+                "name": "comment",
+                "type": "string"
+              },
+              {
+                "id": 2,
+                "name": "environment_id",
+                "type": "string"
+              },
+              {
+                "id": 3,
+                "name": "feature",
+                "type": "feature.Feature"
+              }
+            ]
+          },
+          {
+            "name": "UpdateFeatureResponse"
+          }
+        ],
+        "services": [
+          {
+            "name": "PublicAPIService",
+            "rpcs": [
+              {
+                "name": "GetFeature",
+                "in_type": "GetFeatureRequest",
+                "out_type": "GetFeatureResponse",
+                "options": [
+                  {
+                    "name": "(google.api.http)",
+                    "aggregated": [
+                      {
+                        "name": "get",
+                        "value": "environment/{environment-id}/features/{feature-id}"
+                      }
+                    ]
+                  }
+                ]
+              },
+              {
+                "name": "UpdateFeature",
+                "in_type": "UpdateFeatureRequest",
+                "out_type": "UpdateFeatureResponse",
+                "options": [
+                  {
+                    "name": "(google.api.http)",
+                    "aggregated": [
+                      {
+                        "name": "patch",
+                        "value": "environment/{environment-id}/features/{feature-id}"
+                      }
+                    ]
+                  }
+                ]
+              }
+            ]
+          }
+        ],
+        "imports": [
+          {
+            "path": "google/api/annotations.proto"
+          },
+          {
+            "path": "proto/feature/feature.proto"
+          }
+        ],
+        "package": {
+          "name": "bucketeer.backend"
+        },
+        "options": [
+          {
+            "name": "go_package",
+            "value": "github.com/bucketeer-io/bucketeer/proto/backend"
           }
         ]
       }
