+++ ui/web-v2/apps/admin/src/proto/publicapi/service_pb.d.ts
@@ -1,5 +1,5 @@
 // package: bucketeer.backend
-// file: proto/backend/service.proto
+// file: proto/publicapi/service.proto
 
 import * as jspb from "google-protobuf";
 import * as google_api_annotations_pb from "../../google/api/annotations_pb";
