+++ ui/web-v2/apps/admin/src/proto/publicapi/service_pb.js
@@ -1,4 +1,4 @@
-// source: proto/backend/service.proto
+// source: proto/publicapi/service.proto
 /**
  * @fileoverview
  * @enhanceable
